Added performance output and some cleanup
Now prints the op/s so it can be compared with the C++ version to
measure which is quicker (hoping this should be.). Also, added pop
instruction and removed some temporary junk code as well as added a way
to push variables onto the stack. Need to be able to read variables from
the program memory, still deciding on whether to modify byte-code or not
for the asm interpreter.