Startings of a programming langauge for the VM w/ testing
Decided I wanted to make a programming language for the VM for
educational purposes. It's gonna be quite simple (structs, functions,
strongly typed). I already have the C# version of the emitter for the
virtual machine. I am writting the compiler in C# as I think it will be
quicker to write it in.